+++ server/src/test/java/com/github/dtprj/dongting/raft/store/RetryFrameTest.java
@@ -37,12 +37,12 @@ public class RetryFrameTest extends BaseFiberTest {
     @Test
     public void test() throws Exception {
         // retry and success
-        TestFrame tf = new TestFrame(1, new long[]{1}, false);
+        TestFrame tf = new TestFrame(1, new int[]{1}, false);
         fiberGroup.fireFiber("f", tf);
         tf.future.get(1, TimeUnit.SECONDS);
 
         // fail 2, retry 1
-        tf = new TestFrame(2, new long[]{1}, false);
+        tf = new TestFrame(2, new int[]{1}, false);
         fiberGroup.fireFiber("f", tf);
         try {
             tf.future.get(1, TimeUnit.SECONDS);
@@ -52,12 +52,12 @@ public class RetryFrameTest extends BaseFiberTest {
         }
 
         // fail 2, retry 2
-        tf = new TestFrame(2, new long[]{1, 1}, false);
+        tf = new TestFrame(2, new int[]{1, 1}, false);
         fiberGroup.fireFiber("f", tf);
         tf.future.get(1, TimeUnit.SECONDS);
 
         // fail 3, retry forever
-        tf = new TestFrame(3, new long[]{1, 1}, true);
+        tf = new TestFrame(3, new int[]{1, 1}, true);
         fiberGroup.fireFiber("f", tf);
         tf.future.get(1, TimeUnit.SECONDS);
 
@@ -78,7 +78,7 @@ public class RetryFrameTest extends BaseFiberTest {
 
         // interrupt fiber
         {
-            tf = new TestFrame(1, new long[]{5000}, true);
+            tf = new TestFrame(1, new int[]{5000}, true);
             Fiber fiber = new Fiber("f", fiberGroup, tf);
             fiberGroup.fireFiber(fiber);
             fiberGroup.fireFiber("f2", new FiberFrame<>() {
@@ -98,7 +98,7 @@ public class RetryFrameTest extends BaseFiberTest {
 
         // sub frame throw FiberInterruptException, no retry
         {
-            tf = new TestFrame(1, new long[]{1}, true, new FiberInterruptException("interrupt"));
+            tf = new TestFrame(1, new int[]{1}, true, new FiberInterruptException("interrupt"));
             Fiber fiber = new Fiber("f", fiberGroup, tf);
             fiberGroup.fireFiber(fiber);
             try {
@@ -110,7 +110,7 @@ public class RetryFrameTest extends BaseFiberTest {
         }
 
         // fiber group shutdown, cancel
-        tf = new TestFrame(1, new long[]{5000}, true);
+        tf = new TestFrame(1, new int[]{5000}, true);
         Fiber fiber = new Fiber("f", fiberGroup, tf);
         fiberGroup.fireFiber(fiber);
         shutdownDispatcher();
@@ -125,18 +125,18 @@ public class RetryFrameTest extends BaseFiberTest {
     private static class TestFrame extends FiberFrame<Void> {
         private final CompletableFuture<Void> future = new CompletableFuture<>();
         private final int mockFailCount;
-        private final long[] retryInterval;
+        private final int[] retryInterval;
         private final boolean retryForever;
         private final Exception mockEx;
 
-        public TestFrame(int mockFailCount, long[] retryInterval, boolean retryForever) {
+        public TestFrame(int mockFailCount, int[] retryInterval, boolean retryForever) {
             this.mockFailCount = mockFailCount;
             this.retryInterval = retryInterval;
             this.retryForever = retryForever;
             this.mockEx = new SecurityException("mock error");
         }
 
-        public TestFrame(int mockFailCount, long[] retryInterval, boolean retryForever, Exception mockEx) {
+        public TestFrame(int mockFailCount, int[] retryInterval, boolean retryForever, Exception mockEx) {
             this.mockFailCount = mockFailCount;
             this.retryInterval = retryInterval;
             this.retryForever = retryForever;
